+++ ethergo/submitter/config/iconfig_generated.go
@@ -15,6 +15,8 @@ type IConfig interface {
 	GetBatch(chainID int) bool
 	// GetMaxGasPrice returns the maximum gas price to use for transactions.
 	GetMaxGasPrice(chainID int) (maxPrice *big.Int)
+	// GetBaseGasPrice returns the gas price to be used instead of zero.
+	GetBaseGasPrice(chainID int) (basePrice *big.Int)
 	// GetBumpInterval returns the number of seconds to wait before bumping a transaction
 	// TODO: test this method.
 	GetBumpInterval(chainID int) time.Duration